ipc/client: Warning fix

+++ src/xrt/ipc/client/ipc_client_instance.c
@@ -259,7 +259,7 @@ ipc_connect(struct ipc_connection *ipc_c)
 
 	char sock_file[PATH_MAX];
 
-	int size = u_file_get_path_in_runtime_dir(XRT_IPC_MSG_SOCK_FILENAME, sock_file, PATH_MAX);
+	ssize_t size = u_file_get_path_in_runtime_dir(XRT_IPC_MSG_SOCK_FILENAME, sock_file, PATH_MAX);
 	if (size == -1) {
 		IPC_ERROR(ipc_c, "Could not get socket file name");
 		return false;
